step07/gae-project/requirements.txt
Flask==2.0.1
google-cloud-logging
